requirements.txt
torch>=1.6.0
Pillow~=7.2.0
tqdm~=4.50.2
torchvision>=0.7.0
numpy~=1.19.2
matplotlib~=3.3.2
torchsummary~=1.5.1
piq~=0.5.1
prefetch_generator~=1.0.1
tensorboard~=2.3.0
pytorch_msssim~=0.2.1